chore: refactored endpoints
