+++ Examples/Felis.Router.Test/Program.cs
@@ -2,10 +2,10 @@ using Felis.Router;
 
 var builder = WebApplication.CreateBuilder(args);
 
+builder.Services.AddFelisRouter(builder.Configuration);
+
 var app = builder.Build();
 
 app.UseFelisRouter();
-
-app.MapGet("/", () => "Felis Router is up and running!").ExcludeFromDescription();
 
 app.Run();
